+++ lib/src/views/the_peer_direct_charge_view.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/cupertino.dart';
@@ -8,10 +7,10 @@ import 'package:flutter/services.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 import 'package:thepeer_flutter/src/const/const.dart';
 import 'package:thepeer_flutter/src/model/the_peer_event_model.dart';
-import 'package:thepeer_flutter/src/model/thepeer_success_model.dart';
 import 'package:thepeer_flutter/src/utils/functions.dart';
 import 'package:thepeer_flutter/src/widgets/the_peer_loader.dart';
 import 'package:webview_flutter/webview_flutter.dart';
+import 'package:webview_flutter_android/webview_flutter_android.dart';
 
 import 'package:thepeer_flutter/src/model/thepeer_data.dart';
 import 'package:thepeer_flutter/src/utils/extensions.dart';
@@ -20,7 +19,7 @@ import 'package:thepeer_flutter/src/views/the_peer_error_view.dart';
 
 class ThepeerDirectChargeView extends StatefulWidget {
   /// Public Key from your https://app.withThepeer.com/apps
-  final ThePeerData data;
+  final ThepeerData data;
 
   /// Success callback
   final ValueChanged<dynamic>? onSuccess;
@@ -69,7 +68,7 @@ class ThepeerDirectChargeView extends StatefulWidget {
             topRight: Radius.circular(10),
           ),
           child: SizedBox(
-            height: MediaQuery.of(context).size.height * 0.9,
+            height: MediaQuery.of(context).size.height,
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
@@ -98,8 +97,7 @@ class ThepeerDirectChargeView extends StatefulWidget {
 }
 
 class _ThepeerDirectChargeViewState extends State<ThepeerDirectChargeView> {
-  final _controller = Completer<WebViewController>();
-  Future<WebViewController> get _webViewController => _controller.future;
+  late final WebViewController _controller;
 
   bool _isLoading = true;
   bool get isLoading => _isLoading;
@@ -122,7 +120,7 @@ class _ThepeerDirectChargeViewState extends State<ThepeerDirectChargeView> {
     setState(() {});
   }
 
-  String get createUrl => ThePeerFunctions.createUrl(
+  String get createUrl => ThepeerFunctions.createUrl(
         data: widget.data,
         sdkType: 'directCharge',
       ).toString();
@@ -144,11 +142,11 @@ class _ThepeerDirectChargeViewState extends State<ThepeerDirectChargeView> {
           if (hasError == true) {
             return Center(
               child: widget.errorWidget ??
-                  ThePeerErrorView(
+                  ThepeerErrorView(
                     onClosed: widget.onClosed,
                     reload: () async {
                       setState(() {});
-                      await (await _webViewController).reload();
+                      await _controller.reload();
                     },
                   ),
             );
@@ -167,24 +165,8 @@ class _ThepeerDirectChargeViewState extends State<ThepeerDirectChargeView> {
                 AnimatedOpacity(
                   duration: const Duration(milliseconds: 400),
                   opacity: isLoading == true ? 0 : 1,
-                  child: WebView(
-                    initialUrl: createUrl.toString(),
-                    onWebViewCreated: _controller.complete,
-                    javascriptChannels: _thepeerJavascriptChannel,
-                    javascriptMode: JavascriptMode.unrestricted,
-                    zoomEnabled: false,
-                    onPageStarted: (_) async {
-                      isLoading = true;
-                    },
-                    onWebResourceError: (e) {
-                      hasError = true;
-                      if (widget.showLogs) ThePeerFunctions.log(e.toString());
-                    },
-                    onPageFinished: (_) async {
-                      isLoading = false;
-                      await _injectPeerStack(await _controller.future);
-                    },
-                    navigationDelegate: _handleNavigationInterceptor,
+                  child: WebViewWidget(
+                    controller: _controller,
                   ),
                 ),
               ],
@@ -199,31 +181,11 @@ class _ThepeerDirectChargeViewState extends State<ThepeerDirectChargeView> {
 
   /// Inject JS code to be run in webview
   Future<void> _injectPeerStack(WebViewController controller) {
-    return controller.runJavascript(
-      ThePeerFunctions.peerMessageHandler(
+    return controller.runJavaScript(
+      ThepeerFunctions.peerMessageHandler(
         'ThepeerDirectChargeClientInterface',
       ),
     );
-  }
-
-  /// Javascript channel for events sent by Thepeer
-  Set<JavascriptChannel> get _thepeerJavascriptChannel {
-    return {
-      JavascriptChannel(
-        name: 'ThepeerDirectChargeClientInterface',
-        onMessageReceived: (JavascriptMessage data) {
-          try {
-            if (widget.showLogs)
-              ThePeerFunctions.log('Event: -> ${data.message}');
-            _handleResponse(data.message);
-          } on Exception {
-            if (mounted && widget.onClosed != null) widget.onClosed!();
-          } catch (e) {
-            if (widget.showLogs) ThePeerFunctions.log(e.toString());
-          }
-        },
-      )
-    };
   }
 
   /// Parse event from javascript channel
@@ -246,24 +208,64 @@ class _ThepeerDirectChargeViewState extends State<ThepeerDirectChargeView> {
           return;
       }
     } catch (e) {
-      if (widget.showLogs) ThePeerFunctions.log(e.toString());
+      if (widget.showLogs) ThepeerFunctions.log(e.toString());
     }
   }
 
   /// Handle WebView initialization
   void _handleInit() async {
     await SystemChannels.textInput.invokeMethod<String>('TextInput.hide');
-    if (Platform.isAndroid) WebView.platform = SurfaceAndroidWebView();
+
+    final WebViewController controller =
+        WebViewController.fromPlatformCreationParams(
+            PlatformWebViewControllerCreationParams());
+
+    controller
+      ..enableZoom(false)
+      ..setJavaScriptMode(JavaScriptMode.unrestricted)
+      ..addJavaScriptChannel('ThepeerDirectChargeClientInterface',
+          onMessageReceived: _onMessageReceived)
+      ..setNavigationDelegate(NavigationDelegate(
+          onPageStarted: (_) async {
+            isLoading = true;
+          },
+          onWebResourceError: (e) {
+            hasError = true;
+            if (widget.showLogs) ThepeerFunctions.log(e.toString());
+          },
+          onPageFinished: (_) async {
+            isLoading = false;
+            await _injectPeerStack(_controller);
+          },
+          onNavigationRequest: _handleNavigationInterceptor))
+      ..loadRequest(Uri.parse('$createUrl'));
+    if (controller.platform is AndroidWebViewController) {
+      AndroidWebViewController.enableDebugging(true);
+    }
+    _controller = controller;
+  }
+
+  /// Javascript channel  onMessageRecieved for events sent by Thepeer
+  void _onMessageReceived(JavaScriptMessage data) {
+    try {
+      if (widget.showLogs) ThepeerFunctions.log('Event: -> ${data.message}');
+      _handleResponse(data.message);
+    } on Exception {
+      if (mounted && widget.onClosed != null) widget.onClosed!();
+    } catch (e) {
+      if (widget.showLogs) ThepeerFunctions.log(e.toString());
+    }
   }
 
   NavigationDecision _handleNavigationInterceptor(NavigationRequest request) {
     final url = request.url.toLowerCase();
 
-    if (url.contains('groot.thepeer.co') || url.contains('chain.thepeer.co')) {
+    if (url.contains(ThepeerFunctions.domainName)) {
       // Navigate to all urls contianing Thepeer
       return NavigationDecision.navigate;
     } else {
-      // Block all navigations outside Thepeer
+      //Prevent external navigations from opening in the webview and open in an external browser instead.
+      ThepeerFunctions.launchExternalUrl(url: url, showLogs: (widget.showLogs));
       return NavigationDecision.prevent;
     }
   }