+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [2.0.11]
+Minor changes
+- Redirect external links to the browser
+- Changed widget height to fullscreen for better user experience
+- Removed reference to test/dev domain and created a dynamic base url
+
+Breaking changes
+- Migration from `WebView` to `WebViewWidget`
+- Replaced instances of `ThePeer` with `Thepeer`
+
 ## [2.0.10]
 Bumped up dependency versions